+++ providers/default.rb
@@ -23,8 +23,8 @@ action :uninstall do
   if ::Dir.exist?(new_resource.name)
     ::Dir.open(new_resource.name).each do |dir|
       batch "uninstall #{dir}" do
-        code "\"#{new_resource.name}\\#{dir}\\install\\hdbuninst.exe\" --batch --path=\"#{new_resource.name}\\#{dir}\""
-        only_if { ::File.exist?("#{new_resource.name}\\#{dir}\\install\\hdbuninst.exe") && !dir['hdbclient'].nil? }
+        code "\"#{new_resource.name}/#{dir}/install/hdbuninst.exe\" --batch --path=\"#{new_resource.name}/#{dir}\""
+        only_if { ::File.exist?("#{new_resource.name}/#{dir}/install/hdbuninst.exe") && !dir['hdbclient'].nil? }
       end # end batch
     end # end each
   end # end if
@@ -32,9 +32,10 @@ end # end action
 
 action :install do
   execute "Install HANA Client: #{new_resource.name}" do
-    command "#{new_resource.installer} --batch --path=\"#{new_resource.name}\\hdbclient"
+    cwd ::File.dirname(new_resource.installer)
+    command "#{new_resource.installer} --batch --path=\"#{new_resource.name}/hdbclient\""
     timeout 86_400
     action :run
-    not_if { ::File.exist?("#{new_resource.name}\\hdbclient\\install\\hdbuninst.exe") }
+    not_if { ::File.exist?("#{new_resource.name}/hdbclient/install/hdbuninst.exe") }
   end
 end